Added wind magnification and redirection.
The umbrella calculate an extra wind portion
based on how aligned it is with the wind. It
then redirects this to face the way it is pointing
before adding it to the rest of the wind. Any
wind in the y direction is applied if and only
if the player is already airborne..... the player
cannot be accelerated by the wind if they are already
matching its speed. This may make the umbrella less
fun as it quite limits how much it can move you up.
We may want to make wind y forces also apply when
you are grounded but we need to be wary here since
they lift the player of teh ground disabling
movement intermitently even when exceeded by
gravity.